feat(ui-ux): move contract addresesses config to separate file
